:recycle: Remove too-specific delete and copy actions.
