+++ apps/ui-components/registry/components/node-header/index.tsx
@@ -143,81 +143,6 @@ NodeHeaderAction.displayName = "NodeHeaderAction";
 
 //
 
-export interface NodeHeaderCopyActionProps
-  extends Omit<NodeHeaderActionProps, "onClick"> {
-  onClick?: (nodeId: string, event: React.MouseEvent) => void;
-}
-
-/**
- * A copy action button that passes the node's id to the `onClick` handler when
- * clicked.
- */
-export const NodeHeaderCopyAction = React.forwardRef<
-  HTMLButtonElement,
-  NodeHeaderCopyActionProps
->(({ onClick, ...props }, ref) => {
-  const id = useNodeId();
-
-  const handleClick = useCallback(
-    (event: React.MouseEvent) => {
-      if (!onClick || !id) return;
-
-      onClick(id, event);
-    },
-    [onClick],
-  );
-
-  return (
-    <NodeHeaderAction
-      ref={ref}
-      onClick={handleClick}
-      variant="ghost"
-      {...props}
-    >
-      <Copy />
-    </NodeHeaderAction>
-  );
-});
-
-NodeHeaderCopyAction.displayName = "NodeHeaderCopyAction";
-
-//
-
-export type NodeHeaderDeleteActionProps = Omit<
-  NodeHeaderActionProps,
-  "onClick"
->;
-
-/**
- * A delete action button that removes the node from the graph when clicked.
- */
-export const NodeHeaderDeleteAction = React.forwardRef<
-  HTMLButtonElement,
-  NodeHeaderDeleteActionProps
->((props, ref) => {
-  const id = useNodeId();
-  const { setNodes } = useReactFlow();
-
-  const handleClick = useCallback(() => {
-    setNodes((prevNodes) => prevNodes.filter((node) => node.id !== id));
-  }, []);
-
-  return (
-    <NodeHeaderAction
-      ref={ref}
-      onClick={handleClick}
-      variant="ghost"
-      {...props}
-    >
-      <Trash />
-    </NodeHeaderAction>
-  );
-});
-
-NodeHeaderDeleteAction.displayName = "NodeHeaderDeleteAction";
-
-//
-
 export type NodeHeaderMenuActionProps = Omit<
   NodeHeaderActionProps,
   "onClick"
